requirements.txt
openai==0.10.2
transformers==4.15.0
pytorch==1.8.1
tabulate==0.8.9
pytablewriter==0.62.0
pygments==2.10.0
best-download==0.0.7
datasets==1.11.0
huggingface-hub==0.2.1